requirements.txt
requests>=2.20.0
chardet>=3.0.0
baidu-aip
